Adds keyword constructors and changes order of fields in types
The order of the fields in the Farneback type was changed so that the first field of Farneback and LucasKanade now both represent the number of iterations. This is a breaking change, but brings greater consistency. I've also added constructors with keywords for the fields of the Farneback and LucasKanade type. This commit also includes some demo scripts.